+++ cpp/opendnp3/include/opendnp3/ICommandHandler.h
@@ -41,7 +41,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Select(const ControlRelayOutputBlock& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Select(const ControlRelayOutputBlock& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Operate a ControlRelayOutputBlock - group 12 variation 1
@@ -49,7 +49,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Operate(const ControlRelayOutputBlock& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Operate(const ControlRelayOutputBlock& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Direct operate a ControlRelayOutputBlock - group 12 variation 1
@@ -57,7 +57,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus DirectOperate(const ControlRelayOutputBlock& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus DirectOperate(const ControlRelayOutputBlock& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Select a 16 bit analog output - group 41 variation 2
@@ -65,7 +65,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Select(const AnalogOutputInt16& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Select(const AnalogOutputInt16& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Operate a 16 bit analog output - group 41 variation 2
@@ -73,7 +73,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Operate(const AnalogOutputInt16& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Operate(const AnalogOutputInt16& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Direct operate a 16 bit analog output - group 41 variation 2
@@ -81,7 +81,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus DirectOperate(const AnalogOutputInt16& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus DirectOperate(const AnalogOutputInt16& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Select a 32 bit analog output - group 41 variation 1
@@ -89,7 +89,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Select(const AnalogOutputInt32& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Select(const AnalogOutputInt32& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Operate a 32 bit analog output - group 41 variation 1
@@ -97,7 +97,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Operate(const AnalogOutputInt32& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Operate(const AnalogOutputInt32& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Direct operate a 32 bit analog output - group 41 variation 1
@@ -105,7 +105,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus DirectOperate(const AnalogOutputInt32& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus DirectOperate(const AnalogOutputInt32& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Select a single precision, floating point analog output - group 41 variation 3
@@ -113,7 +113,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Select(const AnalogOutputFloat32& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Select(const AnalogOutputFloat32& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Operate a single precision, floating point analog output - group 41 variation 3
@@ -121,7 +121,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Operate(const AnalogOutputFloat32& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Operate(const AnalogOutputFloat32& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Direct operate a single precision, floating point analog output - group 41 variation 3
@@ -129,7 +129,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus DirectOperate(const AnalogOutputFloat32& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus DirectOperate(const AnalogOutputFloat32& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Select a double precision, floating point analog output - group 41 variation 4
@@ -137,7 +137,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Select(const AnalogOutputDouble64& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Select(const AnalogOutputDouble64& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Operate a double precision, floating point analog output - group 41 variation 4
@@ -145,7 +145,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus Operate(const AnalogOutputDouble64& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus Operate(const AnalogOutputDouble64& arCommand, uint16_t aIndex) = 0;
 
 	/**
 	* Direct operate a double precision, floating point analog output - group 41 variation 4
@@ -153,7 +153,7 @@ public:
 	* @param aIndex index of the command
 	* @return result of request
 	*/
-	virtual CommandStatus DirectOperate(const AnalogOutputDouble64& arCommand, size_t aIndex) = 0;
+	virtual CommandStatus DirectOperate(const AnalogOutputDouble64& arCommand, uint16_t aIndex) = 0;
 
 };
