+++ cpp/opendnp3/include/opendnp3/SimpleCommandHandler.h
@@ -40,25 +40,25 @@ public:
 	*/
 	SimpleCommandHandler(std::function<CommandStatus ()> aStatusFunc);
 
-	CommandStatus Select(const ControlRelayOutputBlock& arCommand, size_t aIndex);
-	CommandStatus Operate(const ControlRelayOutputBlock& arCommand, size_t aIndex);
-	CommandStatus DirectOperate(const ControlRelayOutputBlock& arCommand, size_t aIndex);
+	CommandStatus Select(const ControlRelayOutputBlock& arCommand, uint16_t aIndex);
+	CommandStatus Operate(const ControlRelayOutputBlock& arCommand, uint16_t aIndex);
+	CommandStatus DirectOperate(const ControlRelayOutputBlock& arCommand, uint16_t aIndex);
 
-	CommandStatus Select(const AnalogOutputInt16& arCommand, size_t aIndex);
-	CommandStatus Operate(const AnalogOutputInt16& arCommand, size_t aIndex);
-	CommandStatus DirectOperate(const AnalogOutputInt16& arCommand, size_t aIndex);
+	CommandStatus Select(const AnalogOutputInt16& arCommand, uint16_t aIndex);
+	CommandStatus Operate(const AnalogOutputInt16& arCommand, uint16_t aIndex);
+	CommandStatus DirectOperate(const AnalogOutputInt16& arCommand, uint16_t aIndex);
 
-	CommandStatus Select(const AnalogOutputInt32& arCommand, size_t aIndex);
-	CommandStatus Operate(const AnalogOutputInt32& arCommand, size_t aIndex);
-	CommandStatus DirectOperate(const AnalogOutputInt32& arCommand, size_t aIndex);
+	CommandStatus Select(const AnalogOutputInt32& arCommand, uint16_t aIndex);
+	CommandStatus Operate(const AnalogOutputInt32& arCommand, uint16_t aIndex);
+	CommandStatus DirectOperate(const AnalogOutputInt32& arCommand, uint16_t aIndex);
 
-	CommandStatus Select(const AnalogOutputFloat32& arCommand, size_t aIndex);
-	CommandStatus Operate(const AnalogOutputFloat32& arCommand, size_t aIndex);
-	CommandStatus DirectOperate(const AnalogOutputFloat32& arCommand, size_t aIndex);
+	CommandStatus Select(const AnalogOutputFloat32& arCommand, uint16_t aIndex);
+	CommandStatus Operate(const AnalogOutputFloat32& arCommand, uint16_t aIndex);
+	CommandStatus DirectOperate(const AnalogOutputFloat32& arCommand, uint16_t aIndex);
 
-	CommandStatus Select(const AnalogOutputDouble64& arCommand, size_t aIndex);
-	CommandStatus Operate(const AnalogOutputDouble64& arCommand, size_t aIndex);
-	CommandStatus DirectOperate(const AnalogOutputDouble64& arCommand, size_t aIndex);
+	CommandStatus Select(const AnalogOutputDouble64& arCommand, uint16_t aIndex);
+	CommandStatus Operate(const AnalogOutputDouble64& arCommand, uint16_t aIndex);
+	CommandStatus DirectOperate(const AnalogOutputDouble64& arCommand, uint16_t aIndex);
 
 private:
 	std::function<CommandStatus ()> mStatusFunc;
